requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
ecdsa==0.19.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.7
iso8601==1.1.0
Jinja2==3.1.5
loguru==0.7.2
MarkupSafe==3.0.2
orjson==3.10.7
passlib==1.7.4
pillow==10.4.0
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
pypika-tortoise==0.1.6
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
uvicorn==0.30.6
win32-setctime==1.1.0
